+++ R/2_4_station_list.R
@@ -15,44 +15,61 @@ stationListUI <- function() {
   tagList(
     p(em("Open the panels below to view or download a list of WAV stations, information, and locations shown in this dashboard.")),
     bsCollapse(
-      createStnPanel("Baseline monitoring stations", "baselineDownload", "baselineTable"),
-      createStnPanel("Nutrient monitoring stations", "nutrientDownload", "nutrientTable"),
-      createStnPanel("Thermistor station stations", "thermistorDownload", "thermistorTable"),
-      createStnPanel("Complete station list", "allDownload", "allTable")
+      createStnPanel("Stations currently shown on map", "dl_cur_stns", "cur_stns"),
+      createStnPanel("All baseline monitoring stations", "dl_baseline_stns", "baseline_stns"),
+      createStnPanel("All nutrient monitoring stations", "dl_nutrient_stns", "nutrient_stns"),
+      createStnPanel("All thermistor stations", "dl_therm_stns", "therm_stns"),
+      createStnPanel("All WAV stations", "dl_all_stns", "all_stns")
     )
   )
 }
 
 
 #' requires global data frame 'all_stns'
+#' @param avail_stns list of stations currently shown on the map
 
-stationListServer <- function() {
+stationListServer <- function(stn_list) {
   moduleServer(
     id = "station-list",
     function(input, output, session) {
       ns <- session$ns
 
-      # Defs ----
+      # Station lists ----
 
-      baseline_stns <- filter(all_stns, baseline_stn)
-      nutrient_stns <- filter(all_stns, nutrient_stn)
-      therm_stns <- filter(all_stns, therm_stn)
-
-
-      # Tables ----
-
-      renderStnTbl <- function(df) {
-        renderDataTable({
-          df %>%
-            select(station_id:major_basin) %>%
-            clean_names(case = "big_camel")
-        }, selection = "none")
+      fmtStns <- function(df) {
+        df %>%
+          select(-c(label, baseline_stn, therm_stn, nutrient_stn, map_label)) %>%
+          clean_names(case = "big_camel")
       }
 
-      output$baselineTable <- renderStnTbl(baseline_stns)
-      output$nutrientTable <- renderStnTbl(nutrient_stns)
-      output$thermistorTable <- renderStnTbl(therm_stns)
-      output$allTable <- renderStnTbl(all_stns)
+      cur_stns <- reactive({
+        all_stns %>%
+          filter(station_id %in% stn_list()) %>%
+          fmtStns()
+      })
+
+      baseline_stns <- all_stns %>%
+        filter(baseline_stn) %>%
+        fmtStns()
+
+      nutrient_stns <- all_stns %>%
+        filter(nutrient_stn) %>%
+        fmtStns()
+
+      therm_stns <- all_stns %>%
+        filter(therm_stn) %>%
+        fmtStns()
+
+      all_stns_fmt <- fmtStns(all_stns)
+
+
+      # Output tables ----
+
+      output$cur_stns <- renderDataTable(cur_stns(), selection = "none")
+      output$baseline_stns <- renderDataTable(baseline_stns, selection = "none")
+      output$nutrient_stns <- renderDataTable(nutrient_stns, selection = "none")
+      output$therm_stns <- renderDataTable(therm_stns, selection = "none")
+      output$all_stns <- renderDataTable(all_stns_fmt, selection = "none")
 
 
       # Download handlers ----
@@ -61,10 +78,11 @@ stationListServer <- function() {
         downloadHandler(fname, function(file) { write_csv(df, file) } )
       }
 
-      output$baselineDownload <- createStnDl("wav-baseline-stations.csv", baseline_stns)
-      output$nutrientDownload <- createStnDl("wav-nutrient-stations.csv", nutrient_stns)
-      output$thermistorDownload <- createStnDl("wav-thermistor-stations.csv", therm_stns)
-      output$allDownload <- downloadHandler("wav-station-list.csv", all_stns)
+      output$dl_cur_stns <- createStnDl("WAV Selected Station List.csv", cur_stns())
+      output$dl_baseline_stns <- createStnDl("WAV Baseline Stations.csv", baseline_stns)
+      output$dl_nutrient_stns <- createStnDl("WAV Nutrient Stations.csv", nutrient_stns)
+      output$dl_therm_stns <- createStnDl("WAV Thermistor Stations.csv", therm_stns)
+      output$dl_all_stns <- createStnDl("WAV Station List.csv", all_stns_fmt)
 
     }
   )